Fix OpenJSSE module config (#9316)
* Fix OpenJSSE module config

Avoids

```
java.lang.ExceptionInInitializerError: Exception java.lang.IllegalAccessError: class org.openjsse.sun.security.ssl.SSLLogger (in unnamed
module @0x58cbafc2) cannot access class sun.security.action.GetPropertyAction (in module java.base) because module java.base does not export sun.security.action to
unnamed module @0x58cbafc2 [in thread "Test worker"]
```

* x509

* comment

* better approach

* Avoid test failure

* Cleanup

* Cleanup

+++ okhttp/src/jvmTest/kotlin/okhttp3/CacheTest.kt
@@ -107,7 +107,9 @@ class CacheTest {
   @AfterEach
   fun tearDown() {
     ResponseCache.setDefault(null)
-    cache.delete()
+    if (this::cache.isInitialized) {
+      cache.delete()
+    }
   }
 
   /**
